refactor: cashier p2p and on-ramp pages
