+++ packages/cashier/jest.config.js
@@ -13,6 +13,7 @@ module.exports = {
         '^Components/(.*)$': '<rootDir>/src/components/$1',
         '^Utils/(.*)$': '<rootDir>/src/utils/$1',
         '^Containers/(.*)$': '<rootDir>/src/containers/$1',
+        '^Pages/(.*)$': '<rootDir>/src/pages/$1',
     },
     testPathIgnorePatterns: ['/Routes/', '/validator/'],
     coveragePathIgnorePatterns: [
