+++ packages/cashier/src/containers/index.js
@@ -4,7 +4,5 @@ import Cashier from './cashier.jsx';
 import Deposit from './deposit.jsx';
 import Withdrawal from './withdrawal.jsx';
 import PaymentAgent from './payment-agent.jsx';
-import P2PCashier from './p2p-cashier.jsx';
-import OnRamp from './on-ramp.jsx';
 
-export { Cashier, Deposit, Withdrawal, PaymentAgent, AccountTransfer, PaymentAgentTransfer, P2PCashier, OnRamp };
+export { Cashier, Deposit, Withdrawal, PaymentAgent, AccountTransfer, PaymentAgentTransfer };
