+++ packages/cashier/src/pages/on-ramp/__tests__/on-ramp.spec.js
@@ -19,8 +19,8 @@ jest.mock('@deriv/shared/src/utils/screen/responsive', () => ({
 }));
 jest.mock('Components/Error/cashier-locked', () => () => <div>CashierLocked</div>);
 jest.mock('Components/Error/deposit-locked', () => () => <div>DepositLocked</div>);
-jest.mock('Components/on-ramp-provider-card', () => () => <div>OnRampProviderCard</div>);
-jest.mock('Components/on-ramp-provider-popup', () => () => <div>OnRampProviderPopup</div>);
+jest.mock('Pages/on-ramp/on-ramp-provider-card', () => () => <div>OnRampProviderCard</div>);
+jest.mock('Pages/on-ramp/on-ramp-provider-popup', () => () => <div>OnRampProviderPopup</div>);
 
 describe('<OnRamp />', () => {
     const props = {
