+++ packages/cashier/src/pages/on-ramp/index.js
@@ -1,0 +1,3 @@
+import OnRamp from "./on-ramp";
+
+export default OnRamp;
