+++ packages/cashier/src/pages/on-ramp/on-ramp-provider-card/index.js
@@ -1,0 +1,3 @@
+import OnRampProviderCard from "./on-ramp-provider-card";
+
+export default OnRampProviderCard;
