+++ packages/cashier/src/pages/on-ramp/on-ramp-provider-popup/index.js
@@ -1,0 +1,3 @@
+import OnRampProviderPopup from "./on-ramp-provider-popup";
+
+export default OnRampProviderPopup;
