+++ packages/cashier/src/pages/p2p-cashier/__tests__/p2p-cashier.spec.js
