+++ packages/cashier/src/pages/p2p-cashier/index.js
@@ -1,0 +1,3 @@
+import P2PCashier from "./p2p-cashier";
+
+export default P2PCashier;
