+++ packages/cashier/src/pages/p2p-cashier/p2p-cashier.jsx
